+++ src/compas_viewer/components/renderer/__init__.py
@@ -1,1 +1,1 @@
-from .render import Render  # noqa: F401
+# from .render import Render  # noqa: F401