+++ src/compas_viewer/components/renderer/objects/__init__.py
@@ -1,1 +1,1 @@
-from .object import ViewerObject # noqa: F401
+from .object import ViewerObject  # noqa: F401